refactor(inspection-edit): extract fixed top header into EditorHeader (2487->2087)
Batch 2: moves the fixed top-header inline-JSX block (back link, title, search,
view/batch toggles, progress strip, save status, status badge, theme cycle,
settings, auto-sign, preview, sign, publish) into app/components/editor/
EditorHeader.tsx. The consolidated state object is passed as one prop; 7 closure
values (scheme/setColorScheme/autoSign/handleAutoSignToggle/tenantSlug/
setSignModalOpen/handlePublishClick) are explicit props. Handlers passed by
reference, bodies unchanged.

Behavior-preserving: reviewer diff = MD5-identical JSX modulo the
loaderData.tenantSlug->tenantSlug rename; zero dangling identifiers; no other
route region changed. Gates: type-check 0, lint 0, test:web 516, db:check clean.
